feat: move api switch from the sidebar to the settings
