+++ container/src/features/settings/components/GeneralSettings.tsx
@@ -1,0 +1,44 @@
+import { Label } from '@/components/ui/label';
+import { Input } from '@/components/ui/input';
+import { Button } from '@/components/ui/button';
+import { Separator } from '@/components/ui/separator';
+
+const GeneralSettings = () => {
+  return (
+    <div className="space-y-6 py-4">
+      <div className="grid grid-cols-1 gap-8 lg:grid-cols-3">
+        <div className="lg:col-span-1">
+          <h3 className="text-lg font-medium">Profile</h3>
+          <p className="text-muted-foreground text-sm">Update your personal information.</p>
+        </div>
+        <div className="lg:col-span-2">
+          <div className="space-y-4">
+            <div className="grid grid-cols-1 gap-4 sm:grid-cols-2">
+              <div className="space-y-2">
+                <Label htmlFor="firstName">First Name</Label>
+                <Input id="firstName" placeholder="Enter your first name" className="w-full" />
+              </div>
+              <div className="space-y-2">
+                <Label htmlFor="lastName">Last Name</Label>
+                <Input id="lastName" placeholder="Enter your last name" className="w-full" />
+              </div>
+            </div>
+            <div className="space-y-2">
+              <Label htmlFor="email">Email</Label>
+              <Input id="email" type="email" placeholder="Enter your email" className="w-full max-w-md" />
+            </div>
+          </div>
+        </div>
+      </div>
+      <Separator />
+      <div className="pt-2">
+        <div className="flex gap-3">
+          <Button>Save Changes</Button>
+          <Button variant="outline">Cancel</Button>
+        </div>
+      </div>
+    </div>
+  );
+};
+
+export default GeneralSettings;
