+++ container/src/shared/components/navigation/NavSecondary.tsx
@@ -9,7 +9,6 @@ import {
   SidebarMenuItem,
 } from '@/shared/components/ui/sidebar';
 import SidebarLanguageSwitcher from './SidebarLanguageSwitcher';
-import SidebarApiSwitcher from './SidebarApiSwitcher';
 
 export function NavSecondary({
   items,
@@ -25,7 +24,7 @@ export function NavSecondary({
     <SidebarGroup {...props}>
       <SidebarGroupContent>
         <SidebarMenu>
-          <SidebarApiSwitcher />
+          {/* <SidebarApiSwitcher /> */}
           <SidebarLanguageSwitcher />
           {items.map((item) => (
             <SidebarMenuItem key={item.title}>
